Correctly take preference as a long and stop dividing bounds by E6.

+++ MyTracks/src/com/google/android/apps/mytracks/services/PreferenceManager.java
@@ -130,7 +130,7 @@ public class PreferenceManager {
     }
     if (key == null || key.equals(recording_track_key)) {
       service.setRecordingTrackId(
-          sharedPreferences.getInt(recording_track_key, -1));
+          sharedPreferences.getLong(recording_track_key, -1));
     }
     if (key == null || key.equals(split_frequency_key)) {
       service.getSplitManager().setSplitFrequency(
